# Improve Expression evaluation performance