* Adds support for multi-clause (3+) boolean AND/OR expressions;
* Inlines the `Stack` class to the `Expression.evaluate` method;
* Adds dedicated branches for one and two argument clause functions.
* Replaces the dictionary returned from `Resolver.values` with a lookup function
* Adds errors for accessing `Resolver.single_name` for multi-column cases
* Optimizes the single column access on the `Resolver`.

This reduces the time taken for both getters and setters:

* Getter: single column `is not None` reduced by about 33% (or +50% calls per second)
* Getter: Multi-column expression evaluation is similarly down about -33%
* Setter: about a 10% reduction in time taken compared to before

There's still a difference in performance between a hand-coded `hybrid_property`. Averaging a getter across a True and False result, the hand-coded one takes approximately 0.95 microseconds, whereas the `derived_column` takes closer to 3.7 microseconds. 

Setter amd `expression` performance are still significantly better than on the base `hybrid_property`, with setters being about 3x faster (5us to 1.7us) and the expression over 5x (8us to 1.4us).